Final changes
The map now centres on the chosen location, the weather is also
localised and some comments have been added. the remove location button
works but I would like it to direct back to the main page when its
clicked as well as remove the location. Finally, the nickname of the
location is displayed below the header, this  position is subject to
change.